Exclusão da side bar que estava vinculada a pasta pages e alteração do nome do novo arquivo de interpolação para o app
